requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
djangorestframework==3.14.0
install==1.3.5
mysqlclient==2.2.1
pip==22.0.4
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3.post1
setuptools==58.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3